services/service2/src/encryptednotes/app/requirements.txt
gmpy2==2.1.2
pycryptodome==3.14.1
